styles improved

+++ src/app/app.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Jugador } from '../model/Jugador';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-root',
@@ -27,19 +28,69 @@ export class AppComponent implements OnInit {
   }
 
   resetear(){
-    this.jugadores=[];
+    if(this.jugadores.length>0){
+      Swal.fire({
+        title: 'Seguro?',
+        text: "No se puede deshacer el cambio",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si, resetear!',
+        cancelButtonText: 'Cancelar'
+      }).then((result) => {
+        if (result.value) {   
+  
+          this.jugadores=[];
+          localStorage.setItem("Data",JSON.stringify(this.jugadores));
+          Swal.fire(
+            'Puntos Reseteados',
+            '',
+            'success'
+          )
+        }})
+        .catch();
+
+    }
+   
   }
 
   eliminarJugador(jugador:Jugador){
-   this.jugadores = this.jugadores.filter(j => j._id!=jugador._id);
-   localStorage.setItem("Data",JSON.stringify(this.jugadores));
+
+    Swal.fire({
+      title: 'Seguro?',
+      text: "No se puede deshacer el cambio",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar!',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.value) {   
+        this.jugadores = this.jugadores.filter(j => j._id!=jugador._id);
+        localStorage.setItem("Data",JSON.stringify(this.jugadores));
+      }})
+      .catch();
   }
 
   agregarJugador(){
-    var newJugador= new Jugador(Math.random(),this.newJugador);
-    this.jugadores.push(newJugador);
-    this.newJugador="";
-    localStorage.setItem("Data",JSON.stringify(this.jugadores));
+
+    var jugadorExistente = this.jugadores.filter(j=>j.nombre+""==this.newJugador+"");
+    if(!(jugadorExistente.length>0)){
+      var newJugador= new Jugador(Math.random(),this.newJugador);
+      this.jugadores.push(newJugador);
+      this.newJugador="";
+      localStorage.setItem("Data",JSON.stringify(this.jugadores));
+    }
+    else{
+      Swal.fire(
+        'Jugador Existente',
+        '',
+        'error'
+      )
+    }
+
   }
 
   calcular(jugador:Jugador){
